Refactored getAllItems and getByCategory into a single graphql query. The frontend can now send up an optional category when requesting items.
