+++ frontend/src/components/Items/Add/AddItem.tsx
@@ -1,23 +1,16 @@
 import * as React from 'react';
 
 import { useMutation } from '@apollo/react-hooks';
-import { gql } from 'apollo-boost';
 
 import AddItemView from './AddItemView';
 import { IItem } from '../../../types/item';
 
-const ADD_ITEM_MUTATION = gql`
-  mutation SaveITem($name: String!, $category: String!, $price: String!) {
-    saveItem(name: $name, category: $category, price: $price) {
-      name
-      category
-      price
-    }
-  }
-`;
+import { ADD_ITEM_MUTATION, ITEMS_QUERY } from '../queries';
 
 const AddItem = () => {
-  const [addItem] = useMutation(ADD_ITEM_MUTATION);
+  const [addItem] = useMutation(ADD_ITEM_MUTATION, {
+    refetchQueries: [{ query: ITEMS_QUERY }]
+  });
   const createItem = (item: IItem) => {
     addItem({ variables: { ...item } });
   };
